+++ tbsa-v8m/test_pool/debug/test_d001/secure.c
@@ -147,7 +147,7 @@ void test_payload(tbsa_val_api_t *val)
                     /*Initialize the memory with known data*/
                     val->mem_write((uint32_t *)memory_desc->start, WORD, TEST_DATA);
 
-                    if (test_dbg_seq_write(memory_desc->start, SEQ_OPEN_STATE_READ))
+                    if (test_dbg_seq_write((uint32_t)(memory_desc->start), SEQ_OPEN_STATE_READ))
                         goto clean_up;
 
                     /* Read the data returned by debugger and compare to get the results.*/
@@ -164,7 +164,7 @@ void test_payload(tbsa_val_api_t *val)
                     /*Initialize the memory with known data*/
                     val->mem_write((uint32_t *)memory_desc->start, WORD, ~TEST_DATA);
 
-                    if (test_dbg_seq_write(memory_desc->start, SEQ_OPEN_STATE_WRITE))
+                    if (test_dbg_seq_write((uint32_t)(memory_desc->start), SEQ_OPEN_STATE_WRITE))
                         goto clean_up;
 
                     if (test_dbg_seq_write(TEST_DATA, SEQ_OPEN_STATE_WRITE))
